deprecate: AutoMLHandler() class
