+++ covsirphy/_deprecated/automl_handler.py
@@ -4,11 +4,12 @@
 from datetime import datetime, timedelta
 from itertools import product
 import pandas as pd
+from covsirphy.util.error import deprecate
 from covsirphy.util.validator import Validator
 from covsirphy.util.term import Term
-from covsirphy.automl._mbase import ModelBase
+from covsirphy._deprecated._mbase import ModelBase
 from covsirphy.dynamics.ode import ODEModel
-from covsirphy.automl.autots_predictor import _AutoTSPredictor
+from covsirphy._deprecated.autots_predictor import _AutoTSPredictor
 
 
 class AutoMLHandler(Term):
@@ -37,6 +38,7 @@ class AutoMLHandler(Term):
     _UPPER = "Upper"
     _LOWER = "Lower"
 
+    @deprecate(old="AutoMLHandler()", new="ODEScenario().predict()", version="2.24.0-sigma")
     def __init__(self, X, Y, model, days, **kwargs):
         self._model = Validator(model, "model").subclass(parent=(ModelBase, ODEModel))
         if issubclass(self._model, ModelBase):
