+++ covsirphy/_deprecated/autots_predictor.py
