+++ covsirphy/_deprecated/ode_solver.py
@@ -6,7 +6,7 @@ import pandas as pd
 from scipy.integrate import solve_ivp
 from covsirphy.util.validator import Validator
 from covsirphy.util.term import Term
-from covsirphy.automl._mbase import ModelBase
+from covsirphy._deprecated._mbase import ModelBase
 
 
 class _ODESolver(Term):
