+++ covsirphy/_deprecated/ode_solver_multi.py
@@ -5,7 +5,7 @@ import pandas as pd
 from covsirphy.util.error import NAFoundError
 from covsirphy.util.validator import Validator
 from covsirphy.util.term import Term
-from covsirphy.automl._mbase import ModelBase
+from covsirphy._deprecated._mbase import ModelBase
 from covsirphy._deprecated.ode_solver import _ODESolver
