+++ covsirphy/_deprecated/param_estimator.py
@@ -9,7 +9,7 @@ from covsirphy.util.evaluator import Evaluator
 from covsirphy.util.stopwatch import StopWatch
 from covsirphy.util.validator import Validator
 from covsirphy.util.term import Term
-from covsirphy.automl._mbase import ModelBase
+from covsirphy._deprecated._mbase import ModelBase
 from covsirphy._deprecated.ode_solver import _ODESolver
