+++ covsirphy/_deprecated/phase_estimator.py
@@ -10,7 +10,7 @@ from covsirphy.util.validator import Validator
 from covsirphy.util.term import Term
 from covsirphy._deprecated.jhu_data import JHUData
 from covsirphy._deprecated.population import PopulationData
-from covsirphy.automl._mbase import ModelBase
+from covsirphy._deprecated._mbase import ModelBase
 from covsirphy._deprecated.phase_unit import PhaseUnit
